backend/requirements.txt
fastapi==0.112.2
uvicorn==0.30.6
pydantic==2.9.2
SQLAlchemy==2.0.36
psycopg[binary]==3.2.3
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
pyjwt==2.9.0
httpx==0.27.2
